Add wizard_web_hosting and update install_web_hosting
Install function now expects the new calling interface. It also cooperates with wizard
to get the apache virtual host name. The default is the same as before, the current host name.

The wizard needs more work, it should be configurable and should describe what is going on.
There should be an option not to disable the 000-default site.