Install uwsgi with explicit tarball URL
uWSGI releases on pypi point to uwsgi-latest.tar.gz which is a nice way to get latest uwsgi
if you want to get it quickly. Unfortunately if you want a precise version you need to
install it from tarball directly. Since uwsgi is not a python project really pip does not
know anything about it after installation. The only pip parts are integrationg sufficient
to build and install uwsgi.