Add developer mode.
In developer mode the wsgi file will not use django-seatbelt and will not
constrain how the developer sets up the virtual environment physical layout in
the filesystem.

In the default production mode, seatbelt mandates that all python modules
are in their 'expected' location (that is, setup.py develop in someone's
home folder will not work).