Make cmd_install use bundles instead of requirements file
This rips all the manaul installation code from each install_ step
and changes that to one big "installation" in the install_app stage
where we simply install the bundle we already have from somewhere
else.

This changes the assumptions of some of the previous commands. Most
notably is that requirements file is only used by cmd_bundle and
everything else just does not care.